beamerthemesra: render logos on background not background canvas
Beamer renders a frame with different layers, see figure 1 in the
manual. Previously, the footline and logo was rendered on the background
canvas layer (the deepest one) which makes it impossible to put
something behind it (a background picture). Now it is rendered on the
background layer which allows to put something beneath and above it.

The commit also cleans up the background image. There were some weird
white boxes in it.